+++ template/package.json
@@ -56,7 +56,7 @@
     "react-query-native-devtools": "4.0.0",
     "react-test-renderer": "18.1.0",
     "redux-flipper": "2.0.2",
-    "typescript": "^4.4.4"
+    "typescript": "4.7.4"
   },
   "resolutions": {
     "@types/react": "^17"
