+++ template/src/lib/react-query/hooks/star-wars-hooks.ts
@@ -1,7 +1,8 @@
 import {IStarWars} from '@models/API';
-import {useRQ, UseRQOptions} from '@react-query/custom-hooks';
-import {QueryKey} from '@react-query/query-key';
-import {starWarsService} from '@react-query/service/star-wars';
+
+import {useRQ, UseRQOptions} from '../custom-hooks';
+import {QueryKey} from '../query-key';
+import {starWarsService} from '../service/star-wars';
 
 export function usePeopleList(options?: UseRQOptions<IStarWars.People.Response>) {
   return useRQ([QueryKey.GET_STAR_PEOPLE], starWarsService.people, options);
