+++ template/src/lib/react-query/query-hooks.ts
