+++ template/src/lib/react-query/query-key.ts
