+++ template/src/lib/react-query/query-provider.tsx
