+++ template/src/lib/react-query/service/star-wars.ts
