+++ template/tsconfig.path.json
@@ -4,10 +4,8 @@
       "@router/*": ["./src/router/*"],
       "@models/*": ["./src/model/*"],
       "@config": ["./config/env"],
-      "@redux-store/*": ["./src/redux/*"],
       "@lib/*": ["./src/lib/*"],
       "@i18n": ["./src/i18n"],
-      "@react-query/*": ["./src/react-query/*"],
       "@components-atoms/*": ["./src/components/atoms/*"],
       "@components-molecules/*": ["./src/components/molecules/*"],
       "@components-organisms/*": ["./src/components/organisms/*"],
